BACKEND/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic-settings==2.5.2",
  "pyjwt>=2.10.1",
  "passlib[bcrypt]>=1.7.4",
  "email-validator>=2.0.0",
]

[tool.uv]
# 必要なら設定をここに
